requirements.txt
Flask==3.0.1
Jinja2==3.1.3
pillow==10.2.0
